test(enc): block FMA fusion in the pexp2 golden input to fix arm64 divergence
TestPexp2Golden mismatched cross-arch (arm64 sha 23033e66..., frozen amd64
c7186a45...). Root cause is not pexp2 but the golden's input expression
LCG(&seed)*2110 - 1080: 2110 is not a power of two, so LCG*2110 rounds, and
the compiler fuses the surrounding mul-then-sub into a single FMSUB on arm64
(no FMA under GOAMD64=v1 on amd64). The fused input diverges from the amd64
input for 2132 of the 4096 sweep points, so pexp2 was fed different arguments
on each arch.

Evidence gathered on rpi5 (aarch64, go1.26.1) vs local amd64 (go1.26.5):
for every one of the 1964 sweep inputs whose bits appear on both arches,
pexp2's output bits are identical; a separate 307201-point deterministic
sweep across the underflow-to-zero region (61441 zeros) and the entire
subnormal band (217087 subnormal outputs) hashes to the same sha256 on both
arches. pexp2 is fully cross-arch deterministic, including its subnormal and
underflow tail, and needs no change; math.Ldexp uses the pure-Go path on both
arches (no arch assembly), so its single subnormal multiply is identical.

Wrapping the product in float64() blocks the FMSUB, restoring a
cross-arch-identical input. On amd64 the wrapper is a no-op (fused == blocked,
verified), so the frozen digest c7186a45... is unchanged and now holds on
arm64 too. No re-freeze, no production-code change.

+++ internal/enc/detmath_test.go
@@ -367,7 +367,15 @@ func TestPexp2Golden(t *testing.T) {
 	seed := uint64(99)
 	out := make([]float64, 0, 4096)
 	for range 4096 {
-		out = append(out, pexp2(testsignal.LCG(&seed)*2110-1080))
+		// float64(...) around the product is load-bearing: 2110 is not a
+		// power of two, so LCG*2110 rounds, and a bare LCG*2110 - 1080 fuses
+		// into a single FMSUB on arm64 (no FMA under GOAMD64=v1 on amd64).
+		// That fused input diverges from the amd64 input for ~half the sweep
+		// and is what made this golden mismatch cross-arch; pexp2 itself is
+		// bit-identical on both arches for identical inputs (including the
+		// full subnormal/underflow tail). Blocking the fusion here restores
+		// a cross-arch-identical input, so the frozen digest holds on both.
+		out = append(out, pexp2(float64(testsignal.LCG(&seed)*2110)-1080))
 	}
 	got := sha256Float64s(out...)
 	if pexp2GoldenSHA == "" {
